Sequence optimisation

In MRIBuilder an MR Sequence is defined as a sequence of BuildingBlock objects. Most BuildingBlock objects will contain free parameters determining, for example, the duration of the block or the strength/orientation of the MR gradient. In most MR sequence building software, the user will have to set all of these free parameters by computing the appropriate values given a desired echo time, b-value, etc.

In MRIBuilder the internal free parameters are not set directly. Instead, they are inferred using a non-linear, constrained optimisation. For each sequence type, the developer defines how to compute various summary variables from the BuildingBlock free parameters, such as variables.echo_time, variables.duration, variables.resolution, variables.gradient_strength, variables.diffusion_time, variables.duration_transverse etc. A user can then create a specific instantiation of the sequence by fixing any of these summary variables to their desired values (or setting them to :min/:max to minimise/maximise them). In addition to the user-defined constraints, this optimisation will also take into account any scanner-defined constraints. Internally, MRIBuilder will then optimise the BuildingBlock free parameters to match any user-defined constraints and/or objectives. This optimisation uses the Ipopt optimiser accessed through the JuMP.jl library.

In addition to any user-defined objectives, the developer might also have defined secondary objectives (e.g., minimise the total sequence duration). These objective functions will only be considered if they do not affect the result of the user-defined primary objective. More details on these developer-defined secondary objectives can be found in the section on defining new sequences

Summary variables

All variables are available as members of the variables structure.

MRIBuilder.Variables.variables — Module

Main module containing all the MRIBuilder sequence variables.

All variables are available as members of this module, e.g. variables.echo_time returns the echo time variable. New variables can be defined using @defvar.

Set constraints on variables by passing them on as keywords during the sequence generation, e.g., seq=SpinEcho(echo_time=70).

After sequence generation you can get the variable values by calling variables.echo_time(seq). For the sequence defined above this would return 70. (or a number very close to that).

MRIBuilder.Variables.variables.N_left — Constant

N_left(sinc_pulse)

Number of zero-crossings of the SincPulse before the maximum.

Also, see variables.N_right

MRIBuilder.Variables.variables.N_right — Constant

N_left(sinc_pulse)

Number of zero-crossings of the SincPulse after the maximum.

Also, see variables.N_left

MRIBuilder.Variables.variables.TE — Constant

TE(sequence)

Alternative name to compute the variables.echo_time of a sequence in ms.

MRIBuilder.Variables.variables.TR — Constant

TR(sequence)

Alternative name to compute the variables.repetition_time of a sequence in ms.

MRIBuilder.Variables.variables.all_gradient_strengths — Constant

all_gradient_strengths(spoilt_slice_select)

Returns the gradient strength before, during, and after the pulse in SpoiltSliceSelect.

MRIBuilder.Variables.variables.amplitude — Constant

amplitude(pulse)

Return the amplitude of an RFPulseComponent in kHz.

MRIBuilder.Variables.variables.area_under_curve — Constant

area_under_curve(pathway::Pathway)

Return net displacement in k-space (i.e., spoiling) experienced by the spins following a specific Pathway.

Only gradients active while the spins are in the transverse plane are considered.

Returns a NamedTuple with the area_under_curve for all gradient groups.

MRIBuilder.Variables.variables.bandwidth — Constant

bandwidth(pulse)

Return the bandwidth of an RFPulseComponent in kHz.

MRIBuilder.Variables.variables.bmat — Constant

bmat(pathway::Pathway)

Return 3x3 diffusion-weighted matrix experienced by the spins following a specific Pathway in rad^2 ms/um^2.

Only gradients active while the spins are in the transverse plane are considered.

Returns a NamedTuple with the bmat for all gradient groups.

MRIBuilder.Variables.variables.bmat_gradient — Constant

bmat_gradient(overlapping, qstart[, first_event, last_event])

Computes the addition to the variables.bmat contributed by a specific building block or gradient.

qstart represents the variables.qvec at the start of this component.

If first_event is set to something else than nothing, only the gradient waveform after this RF pulse/Readout will be considered. Similarly, if last_event is set to something else than nothing, only the gradient waveform up to this RF pulse/Readout will be considered.

MRIBuilder.Variables.variables.bval — Constant

bval(pathway::Pathway)

Return size of diffusion-weighting experienced by the spins following a specific Pathway in rad^2 ms/um^2.

Only gradients active while the spins are in the transverse plane will contribute to the diffusion weighting.

Returns a NamedTuple with the bval for all gradient groups.

MRIBuilder.Variables.variables.delay — Constant

delay(sequence)

Returns the offset beetween the readout and the spin echo in ms.

MRIBuilder.Variables.variables.diffusion_time — Constant

diffusion_time(diffusion_sequence)

Returns the diffusion time of a DiffusionSpinEcho in ms.

MRIBuilder.Variables.variables.duration — Constant

duration(block)

Duration of the sequence or building block in ms.

MRIBuilder.Variables.variables.duration_dephase — Constant

duration_dephase(pathway::Pathway)

Returns the net time that spins following the given Pathway spent in the +transverse versus the -transverse state. This determines the amount of T2'-weighting as $e^{t/T_2'}$, where $t$ is the duration_dephase.

Also see variables.duration_transverse for T2-weighting.

MRIBuilder.Variables.variables.duration_state — Constant

duration_state(pathway::Pathway, transverse::Bool, positive::Bool)

Returns how long the Pathway spent in a specific state.

The requested state can be set using transverse and positive as follows:

transverse=false, positive=true: +longitudinal

transverse=true, positive=true: +transverse

transverse=false, positive=false: -longitudinal

transverse=true, positive=false: -transverse

MRIBuilder.Variables.variables.duration_transverse — Constant

duration_transverse(pathway::Pathway)

Returns the total amount of time that spins following the given Pathway spent in the transverse plane. This determines the amount of T2-weighting as $e^{t/T_2}$, where $t$ is the duration_transverse.

Also see variables.duration_dephase for T2'-weighting.

MRIBuilder.Variables.variables.dwell_time — Constant

dwell_time(adc)

The dwell time of the ADC readout in ms.

MRIBuilder.Variables.variables.echo_time — Constant

echo_time(sequence)

Returns the echo time of a sequence in ms.

This is typically defined as the time between the excitation pulse and the crossing of k=0 during the MRI readout.

MRIBuilder.Variables.variables.effective_time — Constant

effective_time(container, indices...)

Returns the start time of component with given indices with respect to the start of the ContainerBlock.

This will crash if the component does not have an variables.effective_time (e.g., if it is (part of) a gradient waveform).

Also see variables.duration, start_time, and end_time

MRIBuilder.Variables.variables.flat_time — Constant

flat_time(trapezoid)

Returns the flat time of a Trapezoid gradient profile in ms.

MRIBuilder.Variables.variables.flip_angle — Constant

flip_angle(pulse)

Return the flip angle of an RFPulseComponent in degrees.

MRIBuilder.Variables.variables.fov — Constant

fov(readout)

Defines the field of view of a readout in mm.

MRIBuilder.Variables.variables.frequency — Constant

frequency(pulse)

Return the off-resonance frequency of an RFPulseComponent in kHz.

MRIBuilder.Variables.variables.gradient_strength — Constant

gradient_strength(gradient)

Maximum 3D gradient strength of the gradient in kHz/um.

MRIBuilder.Variables.variables.gradient_strength_norm — Constant

gradient_strength_norm(gradient)

The norm of the variables.gradient_strength.

MRIBuilder.Variables.variables.lobe_duration — Constant

lobe_duration(sinc_pulse)

Time between two zero-crossings of a SincPulse.

MRIBuilder.Variables.variables.net_dephasing — Constant

net_dephasing(pathway::Pathway)

Return net displacement vector in k-space/q-space experienced by the spins following a specific Pathway in kHz/um.

Only gradients active while the spins are in the transverse plane are considered.

Returns a NamedTuple with the qvec for all gradient groups.

MRIBuilder.Variables.variables.nsamples — Constant

nsamples(adc)

Number of samples in an ADC.

MRIBuilder.Variables.variables.oversample — Constant

oversample(adc)

The oversampling rate of the ADC readout.

MRIBuilder.Variables.variables.phase — Constant

phase(pulse)

Return the phase of an RFPulseComponent in degrees.

MRIBuilder.Variables.variables.qval — Constant

qval(gradient)

The norm of the variables.qvec.

MRIBuilder.Variables.variables.qvec — Constant

qvec(overlapping[, first_event, last_event])

Computes the area under the curve for the gradient waveform in BaseBuildingBlock.

If first_event is set to something else than nothing, only the gradient waveform after this RF pulse/Readout will be considered. Similarly, if last_event is set to something else than nothing, only the gradient waveform up to this RF pulse/Readout will be considered.

MRIBuilder.Variables.variables.ramp_overlap — Constant

ramp_overlap(line_readout)

Return the fraction of the gradient ramp that overlaps with the ADC readout.

Set to 0 to ensure that the ADC is only active during the flat time of the readout.

MRIBuilder.Variables.variables.readout_times — Constant

readout_times(sequence)

Returns all the times that the sequence will readout.

MRIBuilder.Variables.variables.resolution — Constant

resolution(readout)

Resolution of the readout.

MRIBuilder.Variables.variables.rise_time — Constant

rise_time(trapezoid)

Returns the rise time of a Trapezoid gradient profile in ms.

MRIBuilder.Variables.variables.slew_rate — Constant

slew_rate(gradient)

Maximum 3D slew rate of the gradient in kHz/um/ms.

MRIBuilder.Variables.variables.slew_rate_norm — Constant

slew_rate_norm(gradient)

The norm of the variables.slew_rate.

MRIBuilder.Variables.variables.slice_thickness — Constant

slice_thickness(slice_select)

Defines the slice thickness for a RF pulse with an active gradient in mm (e.g., SliceSelect).

Defines as variables.gradient_strength_norm(gradient) / variables.bandwidth(pulse)

MRIBuilder.Variables.variables.spoiler — Constant

spoiler(gradient)

Spatial scale in mm over which the spoiler gradient will dephase by 2π.

Automatically computed based on variables.qvec.

MRIBuilder.Variables.variables.time_to_center — Constant

time_to_center(adc)

The time of the ADC readout to reach the center of k-space.

MRIBuilder.Variables.variables.voxel_size — Constant

voxel_size(readout)

Defines the voxel size of a readout in mm.

MRIBuilder.Variables.variables.Δ — Constant

Δ(sequence)

Alternative name to compute the variables.diffusion_time of a sequence in ms.

MRIBuilder.Variables.variables.δ — Constant

δ(trapezoid)

Returns the effective duration of a Trapezoid gradient profile in ms.

Defined as variables.rise_time + variables.flat_time.

Variables interface

MRIBuilder.Variables — Module

Defines the functions that can be called on parts of an MRI sequence to query or constrain any variables.

In addition this defines:

variables: module containing all variables.

VariableType: parent type for any variables (whether number or JuMP variable).

get_free_variable: helper function to create new JuMP variables.

add_cost_function!: add a specific term to the model cost functions.

set_simple_constraints!: call apply_simple_constraint! for each keyword argument.

apply_simple_constraint!: set a simple equality constraint.

get_pulse/get_gradient/get_readout: Used to get the pulse/gradient/readout part of a building block

gradient_orientation: returns the gradient orientation of a waveform if fixed.

MRIBuilder.Variables.VariableType — Type

Parent type for any variable in the MRI sequence.

Each variable can be one of:

a new JuMP variable

an expression linking this variable to other JuMP variable

a number

Create these using get_free_variable.

MRIBuilder.Variables.variables.TE — Constant

TE(sequence)

Alternative name to compute the variables.echo_time of a sequence in ms.

MRIBuilder.Variables.variables.TR — Constant

TR(sequence)

Alternative name to compute the variables.repetition_time of a sequence in ms.

MRIBuilder.Variables.variables.duration — Constant

duration(block)

Duration of the sequence or building block in ms.

MRIBuilder.Variables.variables.gradient_strength_norm — Constant

gradient_strength_norm(gradient)

The norm of the variables.gradient_strength.

MRIBuilder.Variables.variables.qval — Constant

qval(gradient)

The norm of the variables.qvec.

MRIBuilder.Variables.variables.slew_rate_norm — Constant

slew_rate_norm(gradient)

The norm of the variables.slew_rate.

MRIBuilder.Variables.variables.spoiler — Constant

spoiler(gradient)

Spatial scale in mm over which the spoiler gradient will dephase by 2π.

Automatically computed based on variables.qvec.

MRIBuilder.Variables.variables.Δ — Constant

Δ(sequence)

Alternative name to compute the variables.diffusion_time of a sequence in ms.

MRIBuilder.Variables.AbstractBlock — Type

Parent type of all components, building block, and sequences that form an MRI sequence.

MRIBuilder.Variables.Variable — Type

A sequence property that can be constrained and/or optimised.

It acts as a function, so you can call it on a sequence or building block to get the actual values (e.g., v(sequence)). It can return one of the following:

a number

a vector of number

a NamedTuple with the values for individual sequence components

MRIBuilder.Variables.add_cost_function! — Function

add_cost_function!(function, level=2)

Adds an additional term to the cost function.

This term will be minimised together with any other terms in the cost function. Terms added at a lower level will be optimised before any terms with a higher level.

By default, the term is added to the level=2, which is appropriate for any cost functions added by the developer, which will generally only be optimised after any user-defined cost functions (which are added at level=1 by add_simple_constraint! or set_simple_constraints!.

Any sequence will also have a level=3 cost function, which minimises the total sequence duration.

MRIBuilder.Variables.adjust_groups — Method

adjust_groups(block)

Returns an array of keywords in adjust that should affect a specfic block.

If any of these keywords are present in adjust, then adjust_internal will be called.

Some standard keywords are:

:gradient: expects gradient adjustment parameters

:pulse: expects RF pulse adjustment parameters

MRIBuilder.Variables.adjust_internal — Function

adjust_internal(block, names_used; kwargs...)

Returns the adjusted blocks and add any keywords used in the process to names_used.

This is a helper function used by adjust. It should be defined for any block that is adjustable (as defined by adjust_groups).

MRIBuilder.Variables.apply_simple_constraint! — Method

apply_simple_constraint!(variable, value)

Add a single constraint or objective to the variable.

value can be one of:

nothing: do nothing

:min: minimise the variable

:max: maximise the variable

number: fix variable to this value

equation: fix variable to the result of this equation

apply_simple_constraint!(variable, :>=/:<=, value)

Set an inequality constraint to the variable.

value can be a number of an equation.

MRIBuilder.Variables.base_variables — Method

base_variables([T])

Return dictionary with all Variable objects defined for a specific sequence component/block T.

This only returns those Variable directly defined for this component/block, not for any sub-components (through get_pulse, [get_gradient][(@ref), etc.)

If T is not provided, all Variable objects are returned.

MRIBuilder.Variables.get_free_variable — Method

get_free_variable(value; integer=false, start=0.01)

Get a representation of a given variable given a user-defined constraint.

The result is guaranteed to be a VariableType.

MRIBuilder.Variables.get_readout — Function

get_readout(sequence)

Get the main readout events played out during the sequence.

This has to be defined for individual sequences to work.

Any readout variables not explicitly defined for this sequence will be passed on to the readout.

MRIBuilder.Variables.gradient_orientation — Function

gradient_orientation(building_block)

Returns the gradient orientation.

MRIBuilder.Variables.make_generic — Function

make_generic(sequence/building_block/component)

Returns a generic version of the BaseSequence, BaseBuildingBlock, or BaseComponent

Sequences are all flattened and returned as a single Sequence containing only BuildingBlock objects.

Any BaseBuildingBlock is converted into a BuildingBlock.

Pulses are replaced with GenericPulse (except for instant pulses).

Instant readouts are replaced with ADC.

MRIBuilder.Variables.scanner_constraints! — Method

scanner_constraints!(block)

Constraints variables.gradient_strength and variables.slew_rate to be less than the global_scanner maximum.

MRIBuilder.Variables.set_simple_constraints! — Method

set_simple_constraints!(block, kwargs)

Add any constraints or objective functions to the variables of a AbstractBlock.

Each keyword argument has to match one of the functions in variables(block). If set to a numeric value, a constraint will be added to fix the function value to that numeric value. If set to :min or :max, minimising or maximising this function will be added to the cost function.

MRIBuilder.Variables.variable_defined_for — Method

variable_defined_for(var, Val(type))

Check whether variable is defined for a specific sub-type.

MRIBuilder.Variables.@defvar — Macro

@defvar([getter, ], function(s))

Defines new variables.

Each variable is defined as regular Julia functions embedded within a @defvar macro. For example, to define a variables.echo_time variable for a SpinEcho sequence, one can use:

@defvar echo_time(ge::SpinEcho) = 2 * (variables.effective_time(ge, :refocus) - variables.effective_time(ge, :excitation))

Multiple variables can be defined in a single @defvar by including them in a code block:

@defvar begin
    function var1(seq::SomeSequenceType)
        ...
    end
    function var2(seq::SomeSequenceType)
        ...
    end
end

Before the variable function definitions one can include a getter. This getter defines the type of the sequence component for which the variables will be defined. If the variable is not defined for the sequence, the variable will be extracted for those type of sequence components instead. The following sequence component types are provided:

pulse: use get_pulse

gradient: use get_gradient

readout: use get_readout

pathway: use get_pathway

e.g. the following defines a flip_angle variable, which is marked as a property of an RF pulse.

@defvar pulse flip_angle(...) = ...

If after this definition, flip_angle is not explicitly defined for any sequence, it will be extracted for the RF pulses in that sequence instead.